ubuntu-azure-pipeline: Container deploy
-- An additional check was added to the watcher:
If a kernel is detected, an Azure container will
be deployed to check if the kernel can actually be
installed. If it can be installed, then the validation
job will be triggered. If not, the kernel version will
remain the old one and the next watcher run (after 6
hours) will retry to install the new kernel again.
-- Added Premium storage for Perf
-- Removed 2 Storage TCs
-- Refactored the watcher code